Separate file for onnx provenance tags unit tests
