+++ test/onnx/onnx_import.in.cpp
@@ -356,72 +356,6 @@ NGRAPH_TEST(onnx_${BACKEND_NAME}, model_initializer_wo_input)
     EXPECT_TRUE(test::all_close_f(expected_output, output.front()));
 }
 
-NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_tag_text)
-{
-    const auto function = onnx_import::import_onnx_model(
-        file_util::path_join(SERIALIZED_ZOO, "onnx/provenance_tag_add.prototxt"));
-
-    const auto ng_nodes = function->get_ordered_ops();
-    for (const auto ng_node : ng_nodes)
-    {
-        for (const auto tag : ng_node->get_provenance_tags())
-        {
-            EXPECT_HAS_SUBSTRING(tag, "ONNX");
-        }
-    }
-}
-
-template <typename NodeToCheck>
-void TEST_PROVENANCE_TAGS(const std::string& model_path, const std::string& expected_provenance_tag)
-{
-    const auto function =
-        onnx_import::import_onnx_model(file_util::path_join(SERIALIZED_ZOO, model_path));
-
-    for (const auto ng_node : function->get_ordered_ops())
-    {
-        if (as_type_ptr<NodeToCheck>(ng_node))
-        {
-            const auto tags = ng_node->get_provenance_tags();
-            ASSERT_EQ(tags.size(), 1) << "There should be exactly one provenance tag set for "
-                                      << ng_node;
-
-            EXPECT_EQ(*(tags.cbegin()), expected_provenance_tag);
-        }
-    }
-}
-
-NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_only_outputs)
-{
-    // the Add node in the model does not have a name,
-    // only its output name should be found in the provenance tags
-    TEST_PROVENANCE_TAGS<op::v1::Add>("onnx/provenance_only_outputs.prototxt",
-                                      "<ONNX Add ( -> output_of_add)>");
-}
-
-NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_node_name_and_outputs)
-{
-    TEST_PROVENANCE_TAGS<op::v1::Add>("onnx/provenance_node_name_and_outputs.prototxt",
-                                      "<ONNX Add (Add_node -> output_of_add)>");
-}
-
-NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_multiple_outputs_op)
-{
-    TEST_PROVENANCE_TAGS<op::v1::TopK>("onnx/provenance_multiple_outputs_op.prototxt",
-                                       "<ONNX TopK (TOPK -> values, indices)>");
-}
-
-NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_tagging_constants)
-{
-    TEST_PROVENANCE_TAGS<op::Constant>("onnx/provenance_input_tags.prototxt",
-                                       "<ONNX Input (initializer_of_A)>");
-}
-
-NGRAPH_TEST(onnx_${BACKEND_NAME}, provenance_tagging_parameters)
-{
-    TEST_PROVENANCE_TAGS<op::Parameter>("onnx/provenance_input_tags.prototxt",
-                                        "<ONNX Input (input_B)>");
-}
-
 // ############################################################################ OPERATOR TESTS
 NGRAPH_TEST(onnx_${BACKEND_NAME}, model_addmul_abc)
 {
